src/requirements.txt
Flask~=0.12.2
Flask-Admin~=1.5.0
Flask-SQLAlchemy~=2.2
Flask-Migrate~=2.1.0
Flask-Script~=2.0.5
psycopg2~=2.7.3
cryptography~=2.0.3
